+++ bfabric_app_runner/src/bfabric_app_runner/cli/__main__.py
@@ -15,6 +15,7 @@ from bfabric_app_runner.cli.cmd_action import (
 )
 from bfabric_app_runner.cli.cmd_deploy import cmd_deploy_build_app_zip
 from bfabric_app_runner.cli.cmd_prepare import cmd_prepare_workunit
+from bfabric_app_runner.cli.cmd_run import cmd_run_workunit
 from bfabric_app_runner.cli.inputs import cmd_inputs_prepare, cmd_inputs_clean, cmd_inputs_list, cmd_inputs_check
 from bfabric_app_runner.cli.outputs import cmd_outputs_register, cmd_outputs_register_single_file
 from bfabric_app_runner.cli.validate import (
@@ -78,6 +79,9 @@ cmd_deploy = cyclopts.App(name="deploy", help="Utilities for deploying apps")
 cmd_deploy.command(cmd_deploy_build_app_zip, name="build-app-zip")
 app.command(cmd_deploy)
 
+cmd_run = cyclopts.App(name="run", help="Run an app end-to-end.")
+cmd_run.command(cmd_run_workunit, name="workunit")
+app.command(cmd_run)
 
 if __name__ == "__main__":
     app()