+++ bfabric_app_runner/docs/changelog.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Removed
 
 - AppVersion does not have a `submitter` field anymore.
+- Some old submitter related functionality is deleted!
 
 ## \[0.0.22\] - 2025-05-21